tests/requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-docker==3.2.5
pytest-mock==3.12.0
confluent-kafka==2.6.1
pymysql==1.1.0
python-dotenv==1.0.0
flask==3.1.2
requests==2.32.4